feature: functions to fetch, delete and add exercises to Supabase added

+++ src/Pages/ExercisesList.tsx
@@ -1,12 +1,11 @@
 import * as React from "react";
 import { supabase } from "../supabaseClient";
-
+import { exercisesKey } from "../constants";
 import { useOpen } from "../utils/useOpen";
 import { Exercise } from "../interfaces/exerciseInterfaces";
 import { ItemsList } from "../Components/ItemsList";
 import { AddExerciseForm } from "../Components/AddExerciseForm";
-
-export const exercisesKey = "exercises";
+import { checkUser } from "../utils/checkUser";
 
 export const ExercisesList = () => {
   const { open, setOpen } = useOpen();
@@ -17,58 +16,40 @@ export const ExercisesList = () => {
     getExercises();
   }, []);
 
-  async function getExercises() {
-    const { data } = await supabase.from(exercisesKey).select();
-    setExercises(data as Exercise[]);
-  }
-
-  const handleAddExercise = async (newExercise: Exercise): Promise<void> => {
-    const { data, error } = await supabase
-      .from(exercisesKey)
-      .insert(newExercise)
-      .select();
-
-    if (error) {
-      console.log(error);
-      return;
-    }
-    setExercises((prevExercises) => [
-      ...prevExercises,
-      ...(data as Exercise[]),
-    ]);
-  };
-
-  const handleRemoveExercise = async (id: string): Promise<void> => {
-    try {
-      // Attempt to delete the exercise from Supabase
+  const getExercises = async () => {
+    await checkUser(async (user) => {
       const { data, error } = await supabase
         .from(exercisesKey)
+        .select("*")
+        .eq("user_id", user.id);
+
+      if (error) console.error("Error fetching exercises ", error);
+      else setExercises((data as Exercise[]) || []);
+    });
+  };
+
+  const handleAddExercise = async (newExercise: Exercise): Promise<void> => {
+    await checkUser(async (user) => {
+      const { error } = await supabase
+        .from(exercisesKey)
+        .insert([{ ...newExercise, user_id: user?.id }]);
+
+      if (error) console.error(error);
+      else getExercises();
+    });
+  };
+
+  const handleRemoveExercise = async (exerciseId: string): Promise<void> => {
+    await checkUser(async (user) => {
+      const { error } = await supabase
+        .from("exercises")
         .delete()
-        .eq("id", id)
-        .select();
+        .eq("id", exerciseId)
+        .eq("user_id", user?.id);
 
-      // Log the response data and error
-      console.log("Data:", data);
-      console.log("Error:", error);
-
-      // Check if there was an error during deletion
-      if (error) {
-        console.error("Error deleting exercise:", error.message);
-        return;
-      }
-
-      // If deletion was successful, update local state
-      setExercises((prevExercises) => {
-        const updatedExercises = prevExercises.filter(
-          (exercise) => exercise.id !== id
-        );
-        console.log("Updated Exercises:", updatedExercises);
-        return updatedExercises;
-      });
-    } catch (err) {
-      // Catch any unexpected errors
-      console.error("Unexpected error:", err);
-    }
+      if (error) console.error(error);
+      else getExercises();
+    });
   };
 
   return (
